# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: ms-case-data
  contact: { }
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "8.0"
servers:
  - url: http://localhost:64895
    description: Generated server url
...
paths:
  /{municipalityId}/stakeholders/{stakeholderId}:
    get:
      tags:
        - Stakeholders
      description: Get stakeholder by ID.
      operationId: getStakeholders_1
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StakeholderDTO"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/StakeholderDTO"
    put:
      tags:
        - Stakeholders
      description: Replace stakeholder.
      operationId: putStakeholder
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StakeholderDTO"
        required: true
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content - Successful operation
    patch:
      tags:
        - Stakeholders
      description: Update stakeholder.
      operationId: patchStakeholder
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StakeholderDTO"
        required: true
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content - Successful operation
...
  /{municipalityId}/stakeholders:
    get:
      tags:
        - Stakeholders
      operationId: getStakeholders
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: stakeholderRole
          in: query
          required: false
          schema:
            type: string
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: OK - Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StakeholderDTO"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StakeholderDTO"
...
  /{municipalityId}/errands/{errandId}/stakeholders/{stakeholderId}:
    delete:
      tags:
        - Errands
      description: Delete stakeholder on errand.
      operationId: deleteStakeholder
      parameters:
        - name: municipalityId
          in: path
          required: true
          schema:
            type: string
        - name: errandId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: stakeholderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No content - Successful operation
...
components:
  schemas:
    AddressDTO:
      type: object
      properties:
        addressCategory:
          type: string
          enum:
            - POSTAL_ADDRESS
            - INVOICE_ADDRESS
            - VISITING_ADDRESS
        street:
          maxLength: 255
          minLength: 0
          type: string
          example: Testvägen
        houseNumber:
          maxLength: 255
          minLength: 0
          type: string
          example: "18"
        postalCode:
          maxLength: 255
          minLength: 0
          type: string
          example: 123 45
        city:
          maxLength: 255
          minLength: 0
          type: string
          example: Sundsvall
        country:
          maxLength: 255
          minLength: 0
          type: string
          example: Sverige
        careOf:
          maxLength: 255
          minLength: 0
          type: string
          description: c/o
          example: Test Testorsson
        attention:
          maxLength: 255
          minLength: 0
          type: string
          example: Test Testorsson
        propertyDesignation:
          maxLength: 255
          minLength: 0
          type: string
          example: SUNDSVALL LJUSTA 7:2
        apartmentNumber:
          maxLength: 255
          minLength: 0
          type: string
          example: LGH 1001
        isZoningPlanArea:
          type: boolean
        invoiceMarking:
          maxLength: 255
          minLength: 0
          type: string
          description: "Only in combination with addressCategory: INVOICE_ADDRESS"
        location:
          $ref: "#/components/schemas/CoordinatesDTO"
      description: An stakeholder may have one or more addresses. For example one
        POSTAL_ADDRESS and another INVOICE_ADDRESS.
    ContactInformationDTO:
      type: object
      properties:
        contactType:
          type: string
          enum:
            - CELLPHONE
            - PHONE
            - EMAIL
        value:
          maxLength: 255
          minLength: 0
          type: string
    CoordinatesDTO:
      type: object
      properties:
        latitude:
          type: number
          description: Decimal Degrees (DD)
          format: double
          example: 62.390205
        longitude:
          type: number
          description: Decimal Degrees (DD)
          format: double
          example: 17.306616
    StakeholderDTO:
      required:
        - roles
        - type
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        version:
          type: integer
          format: int32
          readOnly: true
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        type:
          type: string
          enum:
            - PERSON
            - ORGANIZATION
        firstName:
          maxLength: 255
          minLength: 0
          type: string
          example: Test
        lastName:
          maxLength: 255
          minLength: 0
          type: string
          example: Testorsson
        personId:
          type: string
          example: 3ed5bc30-6308-4fd5-a5a7-78d7f96f4438
        organizationName:
          maxLength: 255
          minLength: 0
          type: string
          example: Sundsvalls testfabrik
        organizationNumber:
          maxLength: 255
          minLength: 0
          pattern: "^((18|19|20|21)\\d{6}|\\d{6})-(\\d{4})$"
          type: string
          description: Organization number with 10 or 12 digits.
          example: 19901010-1234
        authorizedSignatory:
          maxLength: 255
          minLength: 0
          type: string
          example: Test Testorsson
        adAccount:
          maxLength: 36
          minLength: 0
          type: string
          description: AD-account
        roles:
          type: array
          description: An stakeholder can have one or more roles.
          items:
            type: string
            description: An stakeholder can have one or more roles.
        addresses:
          type: array
          description: An stakeholder may have one or more addresses. For example
            one POSTAL_ADDRESS and another INVOICE_ADDRESS.
          items:
            $ref: "#/components/schemas/AddressDTO"
        contactInformation:
          type: array
          items:
            $ref: "#/components/schemas/ContactInformationDTO"
        extraParameters:
          type: object
          additionalProperties:
            type: string
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
  securitySchemes: { }